Wait INVITE_TIMEOUT_MS for an m.call.invite
Clients should be inviting the bot to the group chat then immediately
creating a 1:1 room with the bot and sending an m.call.invite to it. If
this doesn't happen after INVITE_TIMEOUT_MS, leave the group chat to
avoid VoIP Conference users persisting in the room when there isn't a
conference going on.